refactor: updating `FloatInRange` to use `validation.FloatBetween` / moving `FloatInSlice`
