+++ internal/services/policy/remediation_management_group.go
@@ -71,7 +71,7 @@ func resourceArmManagementGroupPolicyRemediation() *pluginsdk.Resource {
 			"failure_percentage": {
 				Type:         pluginsdk.TypeFloat,
 				Optional:     true,
-				ValidateFunc: validate2.FloatInRange(0, 1.0),
+				ValidateFunc: validation.FloatBetween(0, 1.0),
 			},
 
 			"parallel_deployments": {
