+++ internal/tf/validation/pluginsdk_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package validate
+package validation
 
 import (
 	"testing"
